refactor things as discussed
